Avoid narrowing conversion error

+++ src/common/mpeg4_p10.cpp
@@ -878,7 +878,7 @@ mpeg4::p10::fix_sps_fps(memory_cptr const &buffer,
       new_avcc.write(nalu);
     }
 
-    return memory_cptr{new memory_c{new_avcc.get_and_lock_buffer(), new_avcc.getFilePointer()}};
+    return memory_cptr{new memory_c{new_avcc.get_and_lock_buffer(), static_cast<size_t>(new_avcc.getFilePointer())}};
 
   } catch(...) {
     return memory_cptr{};
